Refactor database access to use client for queries and remove unused schema file
